Fixed compilation with GCC on Linux
Linkage was failing because apparanetly dlclose@@GLIBC_2.2.5 was
undefined.  This is a symptom that the "linker library" (dl - which
containes the dlclose function) wasn't being linked together with
objects that called dlclose.

Using CMake, this can be resolved by adding ${CMAKE_DL_LIBS} in
target_link_libraries() and hopefully CMake will decide whether or not
the "-ldl" library needs to be included and will set this variable
accordingly.

Tested on Funtoo Linux with GCC 11.03 and it worked.

Added ${CMAKE_DL_LIBS} to target_link_libraries in
src/gui/CMakeLists.txt