Fix names of shared and static libraries
CMake automatically prefixes library names with "lib" on non-Windows
platforms so the library ended up being named "liblibgle-graphics" on
those systems. Fix it by not including the "lib" prefix when calling
add_library.

The change to src/gle/glec.cpp to look for a shared library with the
"lib" prefix only on non-Windows system seems correct but I cannot
verify it because glec.cpp is not compiled by the build system.

Fixes #22