Fix parallel build by not copying gle.cpp anymore
When parallel building, copying gle.cpp to gle_cpp_lib.cpp could happen
twice, once for the static library and once for the shared library, and
this could conflict and cause a build error. The copy was only made so
that different compile definitions could be attached to gle_cpp_lib.cpp.
Instead, attach the compile definitions to the library targets.

Fixes #26

I removed the check for whether HAVE_LIBGLE_STATIC was defined because
nothing in the build system ever defines that macro.